+++ plume/src/test/kotlin/io/github/plume/oss/drivers/NeptuneDriverIntTest.kt
@@ -60,6 +60,7 @@ class NeptuneDriverIntTest {
                 .addHostnames(System.getenv("NEPTUNE_HOSTNAME") ?: "localhost")
                 .port(8182)
                 .keyCertChainFile("src/test/resources/conf/SFSRootCAG2.pem")
+                .clearOnConnect(true)
                 .connect()
         }